backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.2.0
langchain-community==0.2.0
langchain-openai==0.1.8
openai==1.35.0
faiss-cpu>=1.9.0
sentence-transformers==3.0.0
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.3.0
